+++ src/Schema/NamedType/Scalar/Scalar.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/ban-types */
 
+import type { Codec } from './codec.js'
 import { nativeScalarCodecs } from './nativeScalarCodecs.js'
 
 export { nativeScalarCodecs } from './nativeScalarCodecs.js'
@@ -9,15 +10,6 @@ export const ScalarKind = `Scalar`
 export type ScalarKind = typeof ScalarKind
 
 export type StandardScalarRuntimeTypes = boolean | number | string
-
-export const codec = <Decoded = any, Encoded extends StandardScalarRuntimeTypes = StandardScalarRuntimeTypes>(
-  codec: Codec<Decoded, Encoded>,
-) => codec
-
-export type Codec<Decoded = any, Encoded extends StandardScalarRuntimeTypes = StandardScalarRuntimeTypes> = {
-  encode: (value: Decoded) => Encoded
-  decode: (value: Encoded) => Decoded
-}
 
 export const scalar = <$Name extends string, $Codec extends Codec<any, any>>(
   name: $Name,
