+++ src/Schema/NamedType/Scalar/codec.ts
@@ -1,0 +1,10 @@
+import type { StandardScalarRuntimeTypes } from './Scalar.js'
+
+export const codec = <Decoded = any, Encoded extends StandardScalarRuntimeTypes = StandardScalarRuntimeTypes>(
+  codec: Codec<Decoded, Encoded>,
+) => codec
+
+export type Codec<Decoded = any, Encoded extends StandardScalarRuntimeTypes = StandardScalarRuntimeTypes> = {
+  encode: (value: Decoded) => Encoded
+  decode: (value: Encoded) => Decoded
+}
