+++ src/generator/code/scalar.ts
@@ -9,19 +9,12 @@ export const generateScalar = (config: Config) => {
     export * from '${config.libraryPaths.scalars}'
     export * from '${config.importPaths.customScalarCodecs}'
     
-    ${
-    config.typeMapByKind.GraphQLCustomScalarType
-      .map((_) => {
-        return `export type ${_.name} = typeof CustomScalar.${_.name}`
-      }).join(`\n`)
-  }
-
     declare global {
       interface SchemaCustomScalars {
         ${
     config.typeMapByKind.GraphQLCustomScalarType
       .map((_) => {
-        return `${_.name}: ${_.name}`
+        return `${_.name}: typeof CustomScalar.${_.name}`
       }).join(`\n`)
   }
       }
